+++ selene_sdk/samplers/dataloader.py
@@ -3,6 +3,7 @@ This module provides the `SamplerDataLoader` and  `SamplerDataset` classes,
 which allow parallel sampling for any Sampler using
 torch DataLoader mechanism.
 """
+import random
 import  sys
 
 import h5py
@@ -125,6 +126,27 @@ class SamplerDataLoader(DataLoader):
         self.seed = seed
 
 
+def unpackbits_sequence(sequence, s_len):
+    sequence = np.unpackbits(sequence.astype(np.uint8), axis=-2)
+    nulls = np.sum(sequence, axis=-1) == sequence.shape[-1]
+    sequence = sequence.astype(float)
+    sequence[nulls, :] = 1.0 / sequence.shape[-1]
+    if sequence.ndim == 3:
+        sequence = sequence[:, :s_len, :]
+    else:
+        sequence = sequence[:s_len, :]
+    return sequence
+
+
+def unpackbits_targets(targets, t_len):
+    targets = np.unpackbits(targets, axis=-1).astype(float)
+    if targets.ndim == 2:
+        targets = targets[:, :t_len]
+    else:
+        targets = targets[:self.t_len]
+    return targets
+
+
 class _H5Dataset(Dataset):
     """
     This class provides a Dataset that directly loads sequences and targets
@@ -160,13 +182,24 @@ class _H5Dataset(Dataset):
     def __init__(self,
                  file_path,
                  in_memory=False,
-                 unpackbits=False,
+                 unpackbits=False,  # implies unpackbits for both
+                 unpackbits_seq=False,
+                 unpackbits_tgt=False,
                  sequence_key="sequences",
-                 targets_key="targets"):
+                 targets_key="targets",
+                 use_seq_len=None,
+                 shift=None):
         super(_H5Dataset, self).__init__()
         self.file_path = file_path
         self.in_memory = in_memory
+
         self.unpackbits = unpackbits
+        self.unpackbits_seq = unpackbits_seq
+        self.unpackbits_tgt = unpackbits_tgt
+
+        self.use_seq_len = use_seq_len
+        self.shift = shift
+        self._seq_start, self._seq_end = None, None
 
         self._initialized = False
         self._sequence_key = sequence_key
@@ -178,15 +211,22 @@ class _H5Dataset(Dataset):
         def dfunc(self, *args, **kwargs):
             if not self._initialized:
                 self.db = h5py.File(self.file_path, 'r')
+
                 if self.unpackbits:
                     self.s_len = self.db['{0}_length'.format(self._sequence_key)][()]
                     self.t_len = self.db['{0}_length'.format(self._targets_key)][()]
+                elif self.unpackbits_seq:
+                    self.s_len = self.db['{0}_length'.format(self._sequence_key)][()]
+                elif self.unpackbits_tgt:
+                    self.t_len = self.db['{0}_length'.format(self._targets_key)][()]
+
                 if self.in_memory:
                     self.sequences = np.asarray(self.db[self._sequence_key])
                     self.targets = np.asarray(self.db[self._targets_key])
                 else:
                     self.sequences = self.db[self._sequence_key]
                     self.targets = self.db[self._targets_key]
+
                 self._initialized = True
             return func(self, *args, **kwargs)
         return dfunc
@@ -195,25 +235,33 @@ class _H5Dataset(Dataset):
     def __getitem__(self, index):
         if isinstance(index, int):
             index = index % self.sequences.shape[0]
-        sequence = self.sequences[index, :, :]
-        targets = self.targets[index, :]
+        sequence = self.sequences[index]
+        targets = self.targets[index]
+
         if self.unpackbits:
-            sequence = np.unpackbits(sequence, axis=-2)
-            nulls = np.sum(sequence, axis=-1) == sequence.shape[-1]
-            sequence = sequence.astype(float)
-            sequence[nulls, :] = 1.0 / sequence.shape[-1]
-            targets = np.unpackbits(
-                targets, axis=-1).astype(float)
-            if sequence.ndim == 3:
-                sequence = sequence[:, :self.s_len, :]
-            else:
-                sequence = sequence[:self.s_len, :]
-            if targets.ndim == 2:
-                targets = targets[:, :self.t_len]
-            else:
-                targets = targets[:self.t_len]
-        return (torch.from_numpy(sequence.astype(np.float32)),
-                torch.from_numpy(targets.astype(np.float32)))
+            sequence = unpackbits_sequence(sequence, self.s_len)
+            targets = unpackbits_targets(targets, self.t_len)
+        elif self.unpackbits_seq:
+            sequence = unpackbits_sequence(sequence, self.s_len)
+        elif self.unpackbits_tgt:
+            targets = unpackbits_targets(targets, self.t_len)
+
+        if self._seq_start is None:
+            self._seq_start = 0
+            self._seq_end = len(sequence)
+
+            if self.use_seq_len is not None:
+                mid = len(sequence) // 2
+                self._seq_start = int(mid - np.ceil(self.use_seq_len / 2))
+                self._seq_end = mid + self.use_seq_len // 2
+                if self.shift is not None:
+                    self._seq_start += self.shift
+                    self._seq_end += self.shift
+        sequence = sequence[self._seq_start:self._seq_end]
+
+        s = sequence.astype(np.float32)
+        return (torch.from_numpy(s), torch.from_numpy(targets))
+
 
     @init
     def __len__(self):
@@ -288,20 +336,38 @@ class H5DataLoader(DataLoader):
 
     """
     def __init__(self,
-                 filepath,
-                 in_memory=False,
+                 dataset,
                  num_workers=1,
                  use_subset=None,
                  batch_size=1,
-                 shuffle=True,
-                 unpackbits=False,
-                 sequence_key="sequences",
-                 targets_key="targets"):
+                 seed=436,
+                 sampler=None,
+                 batch_sampler=None,
+                 shuffle=True):
+        g = torch.Generator()
+        g.manual_seed(seed)
+
+        def worker_init_fn(worker_id):
+            worker_seed = torch.initial_seed() % 2**32
+            print("Worker seed", worker_seed)
+            np.random.seed(worker_seed)
+            random.seed(worker_seed)
+            torch.manual_seed(worker_seed)
+
         args = {
             "batch_size": batch_size,
-            "num_workers": 0 if in_memory else num_workers,
-            "pin_memory": True
+            "pin_memory": True,
+            "worker_init_fn": worker_init_fn,
+            "sampler": sampler,
+            "batch_sampler": batch_sampler,
+            "generator": g,
         }
+
+        if hasattr(dataset, 'in_memory'):
+            args['num_workers'] = 0 if dataset.in_memory else num_workers
+        else:
+            args['num_workers'] = num_workers
+
         if use_subset is not None:
             from torch.utils.data.sampler import SubsetRandomSampler
             if isinstance(use_subset, int):
@@ -311,10 +377,6 @@ class H5DataLoader(DataLoader):
             args["sampler"] = SubsetRandomSampler(use_subset)
         else:
             args["shuffle"] = shuffle
-        super(H5DataLoader, self).__init__(
-            _H5Dataset(filepath,
-                       in_memory=in_memory,
-                       unpackbits=unpackbits,
-                       sequence_key=sequence_key,
-                       targets_key=targets_key),
-            **args)
+
+        super(H5DataLoader, self).__init__(dataset, **args)
+