+++ selene_sdk/targets/genomic_features.py
@@ -5,8 +5,8 @@ in the sequence.
 
 It accepts the path to a tabix-indexed .bed.gz file of genomic coordinates.
 
-This .tsv/.bed file must contain the following columns, in order:
-    chrom ('1', '2', ..., 'X', 'Y'), start (0-based), end, feature
+This .bed file must contain the following columns, in order:
+    chrom, start (0-based), end, feature (target)
 Additionally, the column names should be omitted from the file itself
 (i.e. there is no header and the first line in the file is the first
 row of genome coordinates for a feature).
@@ -201,10 +201,14 @@ class GenomicFeatures(Target):
     ::
         [chrom, start, end, strand, feature]
 
+    or, if the targets (genomic features) to predict are strand-agnostic:
+    ::
+        [chrom, start, end, feature]
+
 
     Note that `chrom` is interchangeable with any sort of region (e.g.
     a protein in a FAA file). Further, `start` is 0-based. Lastly, any
-    addition columns following the five shown above will be ignored.
+    addition columns following those shown above will be ignored.
 
     Parameters
     ----------
@@ -337,7 +341,7 @@ class GenomicFeatures(Target):
         try:
             tabix_query = self.data.query(chrom, start, end)
             if strand == '+' or strand == '-':
-                return [line for line in tabix_query if str(line[4]) == strand] # strand specificity
+                return [line for line in tabix_query if str(line[3]) == strand] # strand specificity
             else: # not strand specific
                 return tabix_query
         except tabix.TabixError:
@@ -405,7 +409,7 @@ class GenomicFeatures(Target):
             if not rows:
                 return features
             for r in rows:
-                feature = r[3]
+                feature = r[-1]
                 ix = self.feature_index_dict[feature]
                 features[ix] = 1
             return features